# [pytorch][PR] Optimize qavg_pool3d_nhwc