@jamesr66a 2020-03-31
Please revert the autoformatter changes in vec256_qint.h, as they interfere with the blame information